feat: withdraw link from testnet contracts
