+++ packages/hardhat/hardhat.config.ts
@@ -6,6 +6,7 @@ import '@typechain/hardhat';
 import 'hardhat-deploy';
 import 'dotenv/config';
 import { task } from 'hardhat/config';
+import './tasks/withdraw-link';
 import { HardhatUserConfig } from 'hardhat/types';
 
 // This is a sample Hardhat task. To learn how to create your own go to
